Save changes made on transform maps
BUGS:
- The dynamic initializer for the static vector of operations
  is not working, so you cannot run I3T now.

Notes:
I divided the original data map into two private structures
which cannot be swapped as they used to be.
One is an array of bits for marking values that have synergies.
You can query this map for its state by calling
Transformation::getValueState.
The second map is for comfortable validating input values.
It is an array with entries marking values as ZERO, ONE, ANY.