InputManager: Moved mouse button and position update to the beginning of a frame, rather than the end. Fixed isActionTriggered() method.
This fixes an issue with actions that caused InputManager::isActionTriggered(pressed) to always return false if the action was bound to one of the mouse buttons.
The reason was that since mouse buttons were updated at the end of the frame, once a button got pressed, it was still UP during the frame, then at the end of the frame it was set to JUST_DOWN and immediately to DOWN.
The isActionTriggered methods checks for JUST_DOWN key state which mouse buttons never had.

Since the mouse button update was moved it made sense to also move the mouse position update at the begining, this may have adverse effects.
Also some isActionTriggered calls involving the mouse might now actually return true.

Note: Mouse buttons are treated like any other button, and JUST_DOWN is treated as a "click"
But pressing the mouse and starting a drag is NOT considered a click (once the mouse moves some distance), so mouse buttons need special treatment.
One consequence is actions with mouse buttons, isActionTriggered will return true even if the mouse will go on dragging.

Potentially problematic code:
bypassCreateAndPlugConstrutorNodeAction (579)
bypassUnplugAction (599)