Slowing moving functionality from Workspace to DIWNE classes. Separated DiwneObject and Core IDs (Diwne autogenerates IDs now).
Note that interaction is still disabled.
Removed getIdDiwne()
Removed getNodeRectDiwne()
Removed a lot of the node background and header drawing code in Workspace classes. Should be handled by DIWNE in the future.
Created a basic node implementation type that should handle layouting and perhaps pins and collapsing in the future.