Experimenting with layouting. Reworked centering of basic node content. Hopefully will generalize the layout system later.
ImGui tables are lagging behind by one frame when zooming but otherwise work.
Perhaps tables could be replaced with the DiwnePanel layout.
This lagging might be an unavoidable problem I'm not sure yet. Rendering items twice when a change of zoom is detected would likely fix this.
The one frame delay gets particularly bad with low framerate.
Which is notable as currently the node editor seems to be performing worse than it used to => Performance profiling and optimization is needed.
Node must be simplified when zoomed out to reduce vertex counts as sometimes we hit the ImGui drawlist limit, the limit can be raised but that needs further research.