Links can now be replugged by dragging the input pin. They are now colored during creation.
Fixed issue with sequence matrix data being squashed when connecting the matrix input pin by dragging it to an output.
Changed up link colors in general, made the hover indicator thicker and selection be the same color as node selection.
The entire DIWNE styling system is still a bit of a mess intertwined with the I3T themes, this will need a proper solution later.